requirements_test.txt
# Testing Dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-xdist>=3.3.1  # For parallel test execution
pytest-html>=3.2.0   # For HTML reports

# Mocking and HTTP testing
requests-mock>=1.11.0
httpx>=0.24.1

# FastAPI testing
fastapi>=0.100.0
starlette>=0.27.0

# Additional testing utilities
coverage>=7.2.7
tqdm>=4.65.0

# Code quality tools (optional but recommended)
black>=23.7.0        # Code formatting
flake8>=6.0.0        # Linting
mypy>=1.4.1          # Type checking